PEAR-2576 - fix typing on gdc app store hooks
