+++ packages/portal-proto/src/features/annotations/appApi.ts
@@ -11,7 +11,7 @@ const reducers = combineReducers({
 });
 
 export const { id, AppStore, AppContext, useAppSelector, useAppDispatch } =
-  createAppStore({
+  createAppStore<AppState>({
     reducers,
     name: PROJECT_APP_NAME,
     version: "0.0.1",
