+++ packages/portal-proto/src/features/projectsCenter/appApi.ts
@@ -25,7 +25,7 @@ export const reducers = combineReducers({
 });
 
 export const { id, AppStore, AppContext, useAppSelector, useAppDispatch } =
-  createAppStore({
+  createAppStore<AppState>({
     reducers: persistReducer(persistConfig, reducers),
     name: PROJECT_APP_NAME,
     version: "0.0.1",
